Backend/requirements_training.txt
# Training Requirements for Biodiversity Model
tensorflow>=2.13.0
transformers>=4.30.0
librosa>=0.10.1
soundfile>=0.12.1
scikit-maad>=1.4.0
tqdm>=4.65.0
matplotlib>=3.7.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
pillow>=10.0.0